1325 1504 1753 1854 1916 11724 11779 14940 18352 24444 24445 24479 24480 수정

+++ rust/src/01/1325.rs
@@ -11,7 +11,7 @@ fn main() {
     let mut adjacency_list = vec![Vec::new(); n + 1];
 
     for (a, b) in (0..m).map(|_| (input(), input())) {
-        adjacency_list[b as usize].push(a);
+        adjacency_list[b as usize].push(a as i16);
     }
 
     let (counts, max_count) = dfs(&adjacency_list);
@@ -28,14 +28,14 @@ fn main() {
     print!("{output}");
 }
 
-fn dfs(graph: &[Vec<i32>]) -> (Vec<i32>, i32) {
+fn dfs(graph: &[Vec<i16>]) -> (Vec<i16>, i16) {
     let mut counts = vec![0; graph.len()];
     let mut max_count = 0;
 
     for start in 1..graph.len() {
         let mut visited = vec![false; graph.len()];
         let mut count = 0;
-        let mut stack = vec![start as i32];
+        let mut stack = vec![start as i16];
         visited[start] = true;
 
         while let Some(node) = stack.pop() {
