+++ rust/src/24/24444.rs
@@ -4,16 +4,16 @@ use std::io;
 
 fn main() {
     let buf = io::read_to_string(io::stdin()).unwrap();
-    let mut input = buf.split_ascii_whitespace().flat_map(str::parse::<usize>);
+    let mut input = buf.split_ascii_whitespace().flat_map(str::parse::<i32>);
     let mut input = || input.next().unwrap();
     let mut output = String::new();
 
-    let (n, m, r) = (input(), input(), input());
+    let (n, m, r) = (input() as usize, input(), input());
     let mut adjacency_list = vec![Vec::new(); n + 1];
 
     for (u, v) in (0..m).map(|_| (input(), input())) {
-        adjacency_list[u].push(v as i32);
-        adjacency_list[v].push(u as i32);
+        adjacency_list[u as usize].push(v);
+        adjacency_list[v as usize].push(u);
     }
 
     for list in adjacency_list.iter_mut() {
@@ -29,9 +29,9 @@ fn main() {
     print!("{output}");
 }
 
-fn bfs(graph: &[Vec<i32>], start: usize) -> Vec<i32> {
+fn bfs(graph: &[Vec<i32>], start: i32) -> Vec<i32> {
     let mut visited = vec![0; graph.len()];
-    let mut queue = VecDeque::from([start as i32]);
+    let mut queue = VecDeque::from([start]);
     let mut count = 1;
 
     while let Some(node) = queue.pop_front() {
