+++ rust/src/24/24479.rs
@@ -3,16 +3,16 @@ use std::io;
 
 fn main() {
     let buf = io::read_to_string(io::stdin()).unwrap();
-    let mut input = buf.split_ascii_whitespace().flat_map(str::parse::<usize>);
+    let mut input = buf.split_ascii_whitespace().flat_map(str::parse::<i32>);
     let mut input = || input.next().unwrap();
     let mut output = String::new();
 
-    let (n, m, r) = (input(), input(), input());
+    let (n, m, r) = (input() as usize, input(), input());
     let mut adjacency_list = vec![Vec::new(); n + 1];
 
     for (u, v) in (0..m).map(|_| (input(), input())) {
-        adjacency_list[u].push(v as i32);
-        adjacency_list[v].push(u as i32);
+        adjacency_list[u as usize].push(v);
+        adjacency_list[v as usize].push(u);
     }
 
     for list in adjacency_list.iter_mut() {
@@ -28,9 +28,9 @@ fn main() {
     print!("{output}");
 }
 
-fn dfs(graph: &[Vec<i32>], start: usize) -> Vec<i32> {
+fn dfs(graph: &[Vec<i32>], start: i32) -> Vec<i32> {
     let mut visited = vec![0; graph.len()];
-    let mut stack = vec![start as i32];
+    let mut stack = vec![start];
     let mut count = 1;
 
     while let Some(node) = stack.pop() {
